Even bigger changes pls pls pls read
Fleshed out the speech system.  Place a speechblock (prefab found in Main System Speech, totally selfcontained. Child it to the NPC, NOT the spriteobject) on anything that needs to talk.  If it talks by itself, ur done, just enter appropriate ID matching whatevers in the speechrepo.  If it is part of a big group that all says the same thing (mooks, sailors, angry mob, whatever) then simply add a SlaveSpeechModule to that speechblock and  place a master on an empty game object with a trigger + rigidbody.

Speech can be added via SpeechRepo. Just follow the format its in xml